changes:
     * If npm install (or any post install script) fails, the script execution will be terminated. This is useful, since if npm install fails, it likely that any the building process also fails.

+++ setup.sh
@@ -11,6 +11,14 @@ meta git pull
 # install all necessary dependencies
 npm install --no-package-lock
 
+# If npm install (or minstall) fails, stop any further execution.
+# This is advisable since a failed npm install may lead to failures in the
+# building process.
+if [[ "$?" -ne "0" ]]; then
+  printf "\e[1;31mError while executing npm install!\e[0m\n";
+  exit 1;
+fi
+
 # fix conflicting types from jasmine and mocha
 rm -rf node_modules/@types/jasmine
